implementa autorizacao da rota user

+++ src/controllers/user.js
@@ -1,4 +1,4 @@
-const userService = require('../services/user');
+const { userService } = require('../services');
 
 const getAll = async (_req, res, next) => {
     try {
@@ -18,6 +18,7 @@ const getAll = async (_req, res, next) => {
 
 const getById = async (req, res, next) => {
     try {
+        console.log(req.params);
         const { id } = req.params;
 
         const result = await userService.getById(id);
@@ -58,15 +59,21 @@ const update = async (req, res, next) => {
     try {
         const { id } = req.params;
         const { name, description, picture, email, password } = req.body;
+        const { _id: userTokenId, role } = req.user;
 
-        const result = await userService.update(
-            { id, name, description, picture, email, password },
-        );
+        const userFound = await userService.getById(id);
+        
+        const { _id: userFoundId } = userFound;
+
+        if (userFoundId.toString() !== userTokenId && role !== 'admin') {
+            return res.status(401).json({ message: 'Não autorizado.' });
+        }
+        
+        const result = await userService
+            .update({ id, name, description, picture, email, password });
 
         if (!result) {
-            return res.status(400).json({
-                message: 'Erro ao atualizar usuario',
-            });
+            return res.status(400).json({ message: 'Erro ao atualizar usuario' });
         }
 
         return res.status(200).json(result);
@@ -78,6 +85,15 @@ const update = async (req, res, next) => {
 const remove = async (req, res, next) => {
     try {
         const { id } = req.params;
+        const { _id: userTokenId, role } = req.user;
+
+        const userFound = await userService.getById(id);
+        
+        const { _id: userFoundId } = userFound;
+
+        if (userFoundId.toString() !== userTokenId && role !== 'admin') {
+            return res.status(401).json({ message: 'Não autorizado.' });
+        }
 
         const result = await userService.remove(id);
